+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider-utils
 
+## 5.0.0-beta.21
+
+### Patch Changes
+
+- add1126: refactoring: executeTool uses tool as parameter
+
 ## 5.0.0-beta.20
 
 ### Patch Changes