+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,46 @@
 
 ### Patch Changes
 
+- ai@7.0.0-beta.102
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [4873966]
+  - ai@7.0.0-beta.101
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [add1126]
+  - ai@7.0.0-beta.100
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [2a9c144]
+  - ai@7.0.0-beta.99
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@7.0.0-beta.98
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [208d045]
+  - ai@7.0.0-beta.97
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@7.0.0-beta.96
 
 ## 0.0.1-beta.0