+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @ai-sdk/workflow
 
+## 1.0.0-beta.16
+
+### Patch Changes
+
+- ai@7.0.0-beta.102
+
+## 1.0.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies [4873966]
+  - ai@7.0.0-beta.101
+
+## 1.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [add1126]
+  - @ai-sdk/provider-utils@5.0.0-beta.21
+  - ai@7.0.0-beta.100
+
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [2a9c144]
+  - ai@7.0.0-beta.99
+
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- ai@7.0.0-beta.98
+
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [208d045]
+  - ai@7.0.0-beta.97
+
 ## 1.0.0-beta.10
 
 ### Patch Changes